Option to check for pa gapps + data conversion bug fixed

Conflicts:
	src/com/paranoid/paranoidota/updater/server/PaServer.java

+++ src/com/beanstalk/beanstalkota/updater/Server.java
@@ -22,7 +22,7 @@ import com.beanstalk.beanstalkota.updater.Updater.PackageInfo;
 
 public interface Server {
 
-    public String getUrl(String device, double version);
+    public String getUrl(String device, long version);
 
     public List<PackageInfo> createPackageInfoList(String buffer) throws Exception;
